+++ tests/testthat/test-check_cabinet.R
@@ -3,13 +3,10 @@ context("check_cabinet function")
 test_that("finding existing cabinets", {
     withr::with_environment(.GlobalEnv, {
         .test_cab <<- NULL
-        expect_message(check_cabinet(".test_cab"), "Checking cabinet existence...")
+        cli::cli_div(theme = list(".alert-success" = list(before = "PASSING ")))
+        verify_output(test_path("test-find-existing-cabinets.txt"), {
+            check_cabinet(".test_cab")
+        })
         rm(.test_cab, envir = .GlobalEnv)
-    })
-})
-
-test_that("cabinet doesn't exist", {
-    withr::with_environment(.GlobalEnv, {
-        expect_message(check_cabinet(".test_cab"), "Cabinet not found.")
     })
 })