+++ tests/testthat/test-permissions.R
@@ -4,7 +4,10 @@ test_that("permission is granted", {
   options("cabinet.testing" = TRUE)
   options("cabinet.testingPerm" = TRUE)
 
-  expect_message(ask_permission(), "Permission granted.")
+  cli::cli_div(theme = list(".alert-success" = list(before = "PASSING ")))
+  verify_output(test_path("test-permission-granted.txt"), {
+    ask_permission()
+  })
 })
 
 test_that("permission is denied", {
@@ -19,10 +22,20 @@ test_that("check permission catches permission", {
     expect_error(check_permissions())
 
     options("cabinets.permission" = TRUE)
-    expect_message(check_permissions(), "Checking for permissions...")
+
+    cli::cli_div(theme = list(".alert-success" = list(before = "PASSING ")))
+    verify_output(test_path("test-checking-permissions.txt"), {
+      check_permissions()
+    })
 
     options("cabinets.permission" = NULL)
     options("cabinet.testing" = TRUE)
     options("cabinet.testingPerm" = FALSE)
     expect_error(check_permissions())
+})
+
+test_that("permission functions work", {
+  opts <- options()
+  expect_error(perm_no())
+  options(opts)
 })