+++ tests/testthat/test-projects.R
@@ -6,6 +6,11 @@ test_that("project directory already exists", {
     unlink(temp_dir)
 })
 
+skip_if(!cli::is_utf8_output())
+
 test_that("project doesn't exist", {
-    expect_message(check_project(file.path("a", "random", "path")), "Checking if project already exists...")
+    cli::cli_div(theme = list(".alert-success" = list(before = "PASSING ")))
+    verify_output(test_path("test-project-doesnt-exist.txt"), {
+        check_project(file.path("a", "random", "path"))
+    })
 })